fix(svc-external): fix DTOs
- Change KlineDataRequestCoinBase from abstract to concrete class.
- Introduce MexcStatusConverter for better JSON handling of trading pair statuses.
- Update MarketCapUsd property type from int? to long? for improved precision.

+++ src/SVC_Microservices/SVC_External/ApiContracts/Responses/MarketDataProviders/CoinGeckoAssetInfo.cs
@@ -13,7 +13,7 @@ public class CoinGeckoAssetInfo
     /// <summary>
     /// The market capitalization in USD.
     /// </summary>
-    public int? MarketCapUsd { get; set; }
+    public long? MarketCapUsd { get; set; }
 
     /// <summary>
     /// The current price in USD.
